many such changes

+++ companies/models.py
@@ -215,7 +215,7 @@ class Financial(models.Model):
     period_end_date = models.DateField()
     statement = models.CharField(max_length=2, choices=STATEMENT_CHOICES)
     metric = models.ForeignKey(FinancialMetric, on_delete=models.CASCADE, related_name="financials", db_index=False)
-    value = models.DecimalField(max_digits=20, decimal_places=6)
+    value = models.BigIntegerField()
 
     created_at = models.DateTimeField(auto_now_add=True)
